simplify and move logic to compiler crate
