+++ lib/compiler/src/abi.rs
@@ -1,6 +1,7 @@
 //! Shared WebAssembly return-value ABI classification.
 
 use crate::lib::std::vec::Vec;
+use itertools::Itertools;
 use wasmer_types::Type;
 
 /// How a single-register return slot is typed.
@@ -47,4 +48,187 @@ pub enum ReturnAbi {
     Unpacked(Vec<Type>),
     /// Values returned indirectly through a structure-return pointer.
     Sret(Vec<Type>),
+}
+
+/// Two adjacent `F32`s share a vector register (SSE eightbyte); anything
+/// else bit-packs into a raw integer register.
+fn pair_slot(t0: Type, t1: Type) -> PairSlot {
+    if t0 == Type::F32 && t1 == Type::F32 {
+        PairSlot::F32Vector(t0, t1)
+    } else {
+        PairSlot::Raw(t0, t1)
+    }
+}
+
+/// Classifies x86_64 return values.
+pub fn classify_return_type_x86_64(types: &[Type]) -> ReturnAbi {
+    let widths: Vec<_> = types
+        .iter()
+        .map(|ty| match ty {
+            Type::I32 | Type::F32 | Type::ExceptionRef => 32,
+            Type::I64 | Type::F64 => 64,
+            Type::ExternRef | Type::FuncRef => 64,
+            Type::V128 => 128,
+        })
+        .collect_vec();
+
+    match (types, widths.as_slice()) {
+        ([], []) => ReturnAbi::Void,
+        ([value], [_]) => ReturnAbi::Single(*value),
+        ([first, second], [32, 64] | [64, 32] | [64, 64]) => {
+            ReturnAbi::Pair(ReturnSlot::Natural(*first), ReturnSlot::Natural(*second))
+        }
+        ([first, second], [32, 32]) => ReturnAbi::PackedPair(pair_slot(*first, *second)),
+        ([first, second, third], [32, 32, 32 | 64]) => {
+            ReturnAbi::PackedFirst(pair_slot(*first, *second), ReturnSlot::Natural(*third))
+        }
+        ([first, second, third], [64, 32, 32]) => {
+            ReturnAbi::PackedLast(ReturnSlot::Natural(*first), pair_slot(*second, *third))
+        }
+        ([first, second, third, fourth], [32, 32, 32, 32]) => {
+            ReturnAbi::PackedQuads(pair_slot(*first, *second), pair_slot(*third, *fourth))
+        }
+        _ => ReturnAbi::Sret(types.to_vec()),
+    }
+}
+
+/// Classifies AArch64 (AAPCS64) return values.
+///
+/// A float value only gets its own vector register as part of a
+/// homogeneous floating-point aggregate (the `Unpacked` case below).
+pub fn classify_return_type_aarch64(types: &[Type]) -> ReturnAbi {
+    if (2..=4).contains(&types.len())
+        && (types.iter().all(|ty| ty == &Type::F32) || types.iter().all(|ty| ty == &Type::F64))
+    {
+        return ReturnAbi::Unpacked(types.to_vec());
+    }
+
+    if let [first, second] = types
+        && matches!(
+            first,
+            Type::I32 | Type::I64 | Type::F32 | Type::F64 | Type::ExceptionRef
+        )
+        && matches!(second, Type::FuncRef | Type::ExternRef)
+    {
+        return ReturnAbi::Pair(ReturnSlot::Raw(*first), ReturnSlot::Raw(*second));
+    }
+
+    let widths: Vec<_> = types
+        .iter()
+        .map(|ty| match ty {
+            Type::I32 | Type::F32 | Type::ExceptionRef => 32,
+            Type::I64 | Type::F64 | Type::ExternRef | Type::FuncRef => 64,
+            Type::V128 => 128,
+        })
+        .collect_vec();
+
+    match (types, widths.as_slice()) {
+        ([], []) => ReturnAbi::Void,
+        ([value], [_]) => ReturnAbi::Single(*value),
+        ([first, second], [32, 64] | [64, 32] | [64, 64]) => {
+            ReturnAbi::Pair(ReturnSlot::Raw(*first), ReturnSlot::Raw(*second))
+        }
+        ([first, second], [32, 32]) => ReturnAbi::PackedPair(PairSlot::Raw(*first, *second)),
+        ([first, second, third], [32, 32, 32 | 64]) => {
+            ReturnAbi::PackedFirst(PairSlot::Raw(*first, *second), ReturnSlot::Raw(*third))
+        }
+        ([first, second, third], [64, 32, 32]) => {
+            ReturnAbi::PackedLast(ReturnSlot::Raw(*first), PairSlot::Raw(*second, *third))
+        }
+        ([first, second, third, fourth], [32, 32, 32, 32]) => ReturnAbi::PackedQuads(
+            PairSlot::Raw(*first, *second),
+            PairSlot::Raw(*third, *fourth),
+        ),
+        _ => ReturnAbi::Sret(types.to_vec()),
+    }
+}
+
+/// Classifies RISC-V return values.
+pub fn classify_return_type_riscv(types: &[Type], is_riscv64: bool) -> ReturnAbi {
+    let widths: Vec<_> = types
+        .iter()
+        .map(|ty| match ty {
+            Type::I32 | Type::F32 | Type::ExceptionRef => 32,
+            Type::I64 | Type::F64 => 64,
+            Type::ExternRef | Type::FuncRef => {
+                if is_riscv64 {
+                    64
+                } else {
+                    32
+                }
+            }
+            Type::V128 => 128,
+        })
+        .collect_vec();
+
+    match (types, widths.as_slice()) {
+        ([], []) => ReturnAbi::Void,
+        ([value], [_]) => ReturnAbi::Single(*value),
+        ([first, second], [32, 64] | [64, 32] | [64, 64]) => {
+            ReturnAbi::Pair(ReturnSlot::Natural(*first), ReturnSlot::Natural(*second))
+        }
+        ([first, second], [32, 32]) => ReturnAbi::PackedPair(pair_slot(*first, *second)),
+        ([first, second, third], [32, 32, 32 | 64]) => {
+            ReturnAbi::PackedFirst(pair_slot(*first, *second), ReturnSlot::Natural(*third))
+        }
+        ([first, second, third], [64, 32, 32]) => {
+            ReturnAbi::PackedLast(ReturnSlot::Natural(*first), pair_slot(*second, *third))
+        }
+        ([first, second, third, fourth], [32, 32, 32, 32]) => {
+            ReturnAbi::PackedQuads(pair_slot(*first, *second), pair_slot(*third, *fourth))
+        }
+        _ => ReturnAbi::Sret(types.to_vec()),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::classify_return_type_x86_64;
+    use crate::abi::{PairSlot, ReturnAbi, ReturnSlot};
+    use wasmer_types::Type;
+
+    #[test]
+    fn classify_x86_64_return_type_abi() {
+        assert_eq!(classify_return_type_x86_64(&[]), ReturnAbi::Void);
+        assert_eq!(
+            classify_return_type_x86_64(&[Type::I64]),
+            ReturnAbi::Single(Type::I64)
+        );
+        assert_eq!(
+            classify_return_type_x86_64(&[Type::I32, Type::F64]),
+            ReturnAbi::Pair(
+                ReturnSlot::Natural(Type::I32),
+                ReturnSlot::Natural(Type::F64)
+            )
+        );
+        assert_eq!(
+            classify_return_type_x86_64(&[Type::I32, Type::F32]),
+            ReturnAbi::PackedPair(PairSlot::Raw(Type::I32, Type::F32))
+        );
+        assert_eq!(
+            classify_return_type_x86_64(&[Type::F32, Type::F32, Type::I64]),
+            ReturnAbi::PackedFirst(
+                PairSlot::F32Vector(Type::F32, Type::F32),
+                ReturnSlot::Natural(Type::I64)
+            )
+        );
+        assert_eq!(
+            classify_return_type_x86_64(&[Type::F64, Type::I32, Type::F32]),
+            ReturnAbi::PackedLast(
+                ReturnSlot::Natural(Type::F64),
+                PairSlot::Raw(Type::I32, Type::F32)
+            )
+        );
+        assert_eq!(
+            classify_return_type_x86_64(&[Type::I32, Type::F32, Type::F32, Type::I32]),
+            ReturnAbi::PackedQuads(
+                PairSlot::Raw(Type::I32, Type::F32),
+                PairSlot::Raw(Type::F32, Type::I32)
+            )
+        );
+        assert_eq!(
+            classify_return_type_x86_64(&[Type::V128, Type::I32]),
+            ReturnAbi::Sret(vec![Type::V128, Type::I32])
+        );
+    }
 }
